Restructured for DTOs and Service layer
