+++ src/main/java/com/otulp/pluto_payments_backend/Controllers/PaymentController.java
@@ -1,14 +1,15 @@
 package com.otulp.pluto_payments_backend.Controllers;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.google.common.hash.Hashing;
 import com.otulp.pluto_payments_backend.Models.Device;
 import com.otulp.pluto_payments_backend.Models.Payment;
-import com.otulp.pluto_payments_backend.PlutoHelpers.PlutoFormater;
 import com.otulp.pluto_payments_backend.Repositories.DeviceRepository;
+import com.otulp.pluto_payments_backend.Security.HmacChecker;
+import com.otulp.pluto_payments_backend.Services.PaymentService;
+import com.otulp.pluto_payments_backend.Services.impl.PaymentServiceImpl;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -21,22 +22,17 @@ import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
 
 @RestController
+@RequiredArgsConstructor
 public class PaymentController {
 
-    private final long MAXIMUM_TIME_DIFFERENCE_SECONDS = 300;
+    private final PaymentService paymentService;
     private final DeviceRepository deviceRepository;
-
-    public PaymentController(DeviceRepository deviceRepository) {
-        this.deviceRepository = deviceRepository;
-    }
 
     @PostMapping("/authorize")
     public ResponseEntity<String> authorizePayment(
             @RequestBody String rawBody,
             @RequestHeader HttpHeaders headers) {
 
-
-        System.out.println(rawBody);
         ObjectMapper mapper = new ObjectMapper().registerModule(new JavaTimeModule());
 
         Payment payment = null;
@@ -50,13 +46,14 @@ public class PaymentController {
 
         Device device = deviceRepository.findByMacAddress(payment.getDeviceMacAddress());
 
-        if (!PlutoFormater.check_hmac(payment, device, headers, rawBody)) {
+        if (!HmacChecker.checkHmac(payment, device, headers, rawBody)) {
             return ResponseEntity.badRequest().body("HMAC not matching");
         }
 
         LocalDateTime now = LocalDateTime.now();
         long diff = ChronoUnit.SECONDS.between(now, payment.getTimeStamp());
 
+        long MAXIMUM_TIME_DIFFERENCE_SECONDS = 300;
         if (diff > MAXIMUM_TIME_DIFFERENCE_SECONDS) {
             // LOGG
             return ResponseEntity.badRequest().body("Time difference not accepted");
