+++ src/main/java/com/otulp/pluto_payments_backend/Security/HmacChecker.java
@@ -1,16 +1,14 @@
-package com.otulp.pluto_payments_backend.PlutoHelpers;
+package com.otulp.pluto_payments_backend.Security;
 
 import com.google.common.hash.Hashing;
 import com.otulp.pluto_payments_backend.Models.Device;
 import com.otulp.pluto_payments_backend.Models.Payment;
-import org.springframework.http.ResponseEntity;
-
 import org.springframework.http.HttpHeaders;
+
 import java.nio.charset.StandardCharsets;
 
-public class PlutoFormater {
-
-    public static boolean check_hmac(Payment payment, Device device, HttpHeaders headers, String rawBody) {
+public class HmacChecker {
+    public static boolean checkHmac(Payment payment, Device device, HttpHeaders headers, String rawBody) {
         // 1. Create and compare hmac
         String hmac = headers.getFirst("Authorization");
         String url = "http://" + headers.getFirst("Host") + "/authorize"; // <-------------------- CHANGE HERE WHEN GOING FOR HTTPS
@@ -27,8 +25,6 @@ public class PlutoFormater {
                 url + "\n" +
                 content + "\n" +
                 bodyHex;
-
-
 
         canonicalString += device.getDeviceKey();
