+++ src/main/java/com/otulp/pluto_payments_backend/Services/PaymentService.java
@@ -1,0 +1,8 @@
+package com.otulp.pluto_payments_backend.Services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public interface PaymentService {
+
+}
