+++ src/main/java/com/otulp/pluto_payments_backend/Services/impl/PaymentServiceImpl.java
@@ -1,0 +1,12 @@
+package com.otulp.pluto_payments_backend.Services.impl;
+
+import com.google.common.hash.Hashing;
+import com.otulp.pluto_payments_backend.Models.Device;
+import com.otulp.pluto_payments_backend.Models.Payment;
+import com.otulp.pluto_payments_backend.Services.PaymentService;
+import org.springframework.http.HttpHeaders;
+
+import java.nio.charset.StandardCharsets;
+
+public class PaymentServiceImpl implements PaymentService {
+}
